+++ src/lib.rs
@@ -92,11 +92,13 @@ pub mod http;
 pub mod player;
 pub mod protocol;
 pub mod proxy;
-pub mod rand;
 pub mod remote;
 pub mod tokens;
 pub mod track;
 pub mod util;
+pub mod uuid;
 
 #[allow(clippy::enum_glob_use)]
 pub mod error;
+
+pub use uuid::Uuid;