+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
 ### Added
 - [docs] Comprehensive documentation for all public APIs and internals
+- [uuid] `uuid` module providing a fast UUID v4 generator
 
 ## [0.3.0] - 2024-11-28